Test for IP viscosity.
This test verifies that the Interior Penalty (IP) method for viscosity
converges at the expected order.  The viscosity term is of the form:
div(nu [(grad u)+(grad u)^T]) with variable viscosity nu. Convergence is
tested for RT2, DG1, BDM1 and BDM2 vector fields.

This also serves as a regression test against coffee issues on the bendy
branch with the internal facet integrals present in the IP scheme.