Use PETSc attribute writing to store metadata
Only possible for int-valued attributes for now.  Additionally add
optional functionality to convert the Viewer file handle to an h5py.File
object.  This latter only works if PETSc and h5py were linked to the
same HDF5 library, since this is somewhat complex to get right, we don't
rely on it.